Working on moving tool modes to behaviors.
