+++ InDappledGroves/Items/ItemIDGAxe.cs
@@ -22,34 +22,34 @@ namespace InDappledGroves.Items.Tools
         {
             base.OnLoaded(api);
             ICoreClientAPI capi = api as ICoreClientAPI;
-            this.toolModes = ObjectCacheUtil.GetOrCreate<SkillItem[]>(api, "idgAxeToolModes", delegate
-            {
-                SkillItem[] array;
-                array = new SkillItem[]
-                {
-                        new SkillItem
-                        {
-                            Code = new AssetLocation("chopping"),
-                            Name = Lang.Get("Chopping", Array.Empty<object>())
-                        },
-                        new SkillItem
-                        {
-                            Code = new AssetLocation("planing"),
-                            Name = Lang.Get("Planing", Array.Empty<object>())
-                        }
-                };
-                if (capi != null)
-                {
-                    array[0].WithIcon(capi, capi.Gui.LoadSvgWithPadding(new AssetLocation("indappledgroves:textures/icons/axechop.svg"), 48, 48, 5, new int?(-1)));
-                    array[0].TexturePremultipliedAlpha = false;
-                    if (array.Length > 1)
-                    {
-                        array[1].WithIcon(capi, capi.Gui.LoadSvgWithPadding(new AssetLocation("indappledgroves:textures/icons/axestrip.svg"), 48, 48, 5, new int?(-1)));
-                        array[1].TexturePremultipliedAlpha = false;
-                    }
-                }
-                return array;
-            });
+            //this.toolModes = ObjectCacheUtil.GetOrCreate<SkillItem[]>(api, "idgAxeToolModes", delegate
+            //{
+            //    SkillItem[] array;
+            //    array = new SkillItem[]
+            //    {
+            //            new SkillItem
+            //            {
+            //                Code = new AssetLocation("chopping"),
+            //                Name = Lang.Get("Chopping", Array.Empty<object>())
+            //            },
+            //            new SkillItem
+            //            {
+            //                Code = new AssetLocation("planing"),
+            //                Name = Lang.Get("Planing", Array.Empty<object>())
+            //            }
+            //    };
+            //    if (capi != null)
+            //    {
+            //        array[0].WithIcon(capi, capi.Gui.LoadSvgWithPadding(new AssetLocation("indappledgroves:textures/icons/axechop.svg"), 48, 48, 5, new int?(-1)));
+            //        array[0].TexturePremultipliedAlpha = false;
+            //        if (array.Length > 1)
+            //        {
+            //            array[1].WithIcon(capi, capi.Gui.LoadSvgWithPadding(new AssetLocation("indappledgroves:textures/icons/axestrip.svg"), 48, 48, 5, new int?(-1)));
+            //            array[1].TexturePremultipliedAlpha = false;
+            //        }
+            //    }
+            //    return array;
+            //});
         }
 
         static ItemIDGAxe()
@@ -94,25 +94,25 @@ namespace InDappledGroves.Items.Tools
 
             return base.OnBlockBreaking(player, blockSel, itemslot, remainingResistance, dt / treeResistance, counter);
         }
-        #region ToolMode Stuff
-        // Token: 0x06001847 RID: 6215 RVA: 0x000E49D4 File Offset: 0x000E2BD4
-        public override SkillItem[] GetToolModes(ItemSlot slot, IClientPlayer forPlayer, BlockSelection blockSel)
-        {
-            return this.toolModes;
-        }
+        //#region ToolMode Stuff
+        //// Token: 0x06001847 RID: 6215 RVA: 0x000E49D4 File Offset: 0x000E2BD4
+        //public override SkillItem[] GetToolModes(ItemSlot slot, IClientPlayer forPlayer, BlockSelection blockSel)
+        //{
+        //    return this.toolModes;
+        //}
 
-        // Token: 0x06001848 RID: 6216 RVA: 0x000E49DC File Offset: 0x000E2BDC
-        public override int GetToolMode(ItemSlot slot, IPlayer byPlayer, BlockSelection blockSel)
-        {
-            return Math.Min(this.toolModes.Length - 1, slot.Itemstack.Attributes.GetInt("toolMode", 0));
-        }
+        //// Token: 0x06001848 RID: 6216 RVA: 0x000E49DC File Offset: 0x000E2BDC
+        //public override int GetToolMode(ItemSlot slot, IPlayer byPlayer, BlockSelection blockSel)
+        //{
+        //    return Math.Min(this.toolModes.Length - 1, slot.Itemstack.Attributes.GetInt("toolMode", 0));
+        //}
 
-        // Token: 0x06001849 RID: 6217 RVA: 0x000C8EF1 File Offset: 0x000C70F1
-        public override void SetToolMode(ItemSlot slot, IPlayer byPlayer, BlockSelection blockSel, int toolMode)
-        {
-            slot.Itemstack.Attributes.SetInt("toolMode", toolMode);
-        }
-        #endregion ToolMode Stuff
+        //// Token: 0x06001849 RID: 6217 RVA: 0x000C8EF1 File Offset: 0x000C70F1
+        //public override void SetToolMode(ItemSlot slot, IPlayer byPlayer, BlockSelection blockSel, int toolMode)
+        //{
+        //    slot.Itemstack.Attributes.SetInt("toolMode", toolMode);
+        //}
+        //#endregion ToolMode Stuff
         public override bool OnBlockBrokenWith(IWorldAccessor world, Entity byEntity, ItemSlot itemslot, BlockSelection blockSel, float dropQuantityMultiplier = 1)
         {
             IPlayer byPlayer = null;
